requirements.txt
tk
Pillow==9.1.0
